+++ inventory_management_system_api/routers/v1/usage_status.py
@@ -3,16 +3,23 @@ Module for providing an API router which defines routes for managing Usage statu
 service.
 """
 
+# We don't define docstrings in router methods as they would end up in the openapi/swagger docs. We also expect
+# some duplicate code inside routers as the code is similar between entities and error handling may be repeated.
+# pylint: disable=missing-function-docstring
+# pylint: disable=duplicate-code
+
 import logging
 from typing import Annotated
 
 from fastapi import APIRouter, Depends, HTTPException, Path, status
 
+from inventory_management_system_api.auth.authorisation import AuthorisedDep
 from inventory_management_system_api.core.exceptions import (
     DuplicateRecordError,
     InvalidObjectIdError,
     MissingRecordError,
     PartOfItemError,
+    PartOfRuleError,
 )
 from inventory_management_system_api.schemas.usage_status import UsageStatusPostSchema, UsageStatusSchema
 from inventory_management_system_api.services.usage_status import UsageStatusService
@@ -31,11 +38,14 @@ UsageStatusServiceDep = Annotated[UsageStatusService, Depends(UsageStatusService
     status_code=status.HTTP_201_CREATED,
 )
 def create_usage_status(
-    usage_status: UsageStatusPostSchema, usage_status_service: UsageStatusServiceDep
+    usage_status: UsageStatusPostSchema, usage_status_service: UsageStatusServiceDep, authorised: AuthorisedDep
 ) -> UsageStatusSchema:
-    # pylint: disable=missing-function-docstring
     logger.info("Creating a new usage status")
     logger.debug("Usage status data: %s", usage_status)
+
+    # check user is authorised to perform operation
+    if not authorised:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorised to perform this operation")
 
     try:
         usage_status = usage_status_service.create(usage_status)
@@ -47,6 +57,14 @@ def create_usage_status(
         raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail=message) from exc
 
 
+@router.get(path="", summary="Get usage statuses", response_description="List of usage statuses")
+def get_usage_statuses(usage_status_service: UsageStatusServiceDep) -> list[UsageStatusSchema]:
+    logger.info("Getting Usage statuses")
+
+    usage_statuses = usage_status_service.list()
+    return [UsageStatusSchema(**usage_status.model_dump()) for usage_status in usage_statuses]
+
+
 @router.get(
     path="/{usage_status_id}",
     summary="Get a usage status by ID",
@@ -56,8 +74,7 @@ def get_usage_status(
     usage_status_id: Annotated[str, Path(description="The ID of the usage status to be retrieved")],
     usage_status_service: UsageStatusServiceDep,
 ) -> UsageStatusSchema:
-    # pylint: disable=missing-function-docstring
-    logger.info("Getting usage status with ID %s", usage_status_id)
+    logger.info("Getting usage status with ID '%s'", usage_status_id)
     message = "Usage status not found"
     try:
         usage_status = usage_status_service.get(usage_status_id)
@@ -70,15 +87,6 @@ def get_usage_status(
     return UsageStatusSchema(**usage_status.model_dump())
 
 
-@router.get(path="", summary="Get usage statuses", response_description="List of usage statuses")
-def get_usage_statuses(usage_status_service: UsageStatusServiceDep) -> list[UsageStatusSchema]:
-    # pylint: disable=missing-function-docstring
-    logger.info("Getting Usage statuses")
-
-    usage_statuses = usage_status_service.list()
-    return [UsageStatusSchema(**usage_status.model_dump()) for usage_status in usage_statuses]
-
-
 @router.delete(
     path="/{usage_status_id}",
     summary="Delete a usage status by its ID",
@@ -88,9 +96,14 @@ def get_usage_statuses(usage_status_service: UsageStatusServiceDep) -> list[Usag
 def delete_usage_status(
     usage_status_id: Annotated[str, Path(description="ID of the usage status to delete")],
     usage_status_service: UsageStatusServiceDep,
+    authorised: AuthorisedDep,
 ) -> None:
-    # pylint: disable=missing-function-docstring
-    logger.info("Deleting usage status with ID: %s", usage_status_id)
+    logger.info("Deleting usage status with ID '%s'", usage_status_id)
+
+    # check user is authorised to perform operation
+    if not authorised:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorised to perform this operation")
+
     try:
         usage_status_service.delete(usage_status_id)
     except (MissingRecordError, InvalidObjectIdError) as exc:
@@ -98,6 +111,10 @@ def delete_usage_status(
         logger.exception(message)
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=message) from exc
     except PartOfItemError as exc:
-        message = "The specified usage status is part of an Item"
+        message = "The specified usage status is part of an item"
+        logger.exception(message)
+        raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail=message) from exc
+    except PartOfRuleError as exc:
+        message = "The specified usage status is part of a rule"
         logger.exception(message)
         raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail=message) from exc
